Core: fix data type for static files length
This patch makes Monkey behave right when handling big files.

+++ src/include/mk_plugin.h
@@ -106,7 +106,7 @@ struct plugin_network_io
     int (*writev) (int, struct mk_iov *);
     int (*close) (int);
     int (*connect) (char *, int);
-    int (*send_file) (int, int, off_t *, size_t);
+    ssize_t (*send_file) (int, int, off_t *, size_t);
     int (*create_socket) (int, int, int);
     int (*bind) (int, const struct sockaddr *addr, socklen_t, int);
     int (*server) (int, char *, int);
@@ -185,7 +185,7 @@ struct plugin_api
     int (*socket_sendv) (int, struct mk_iov *);
     int (*socket_send) (int, const void *, size_t);
     int (*socket_read) (int, void *, int);
-    int (*socket_send_file) (int, int, off_t *, size_t);
+    ssize_t (*socket_send_file) (int, int, off_t *, size_t);
     int (*socket_ip_str) (int, char **, int, unsigned long *);
 
     struct server_config *config;
